split delete core_instance cmd
