+++ cmd/calyptia/delete_core_instance_aws.go
@@ -1,0 +1,21 @@
+package main
+
+import (
+	"errors"
+
+	"github.com/spf13/cobra"
+)
+
+func newCmdDeleteCoreInstanceOnAWS(config *config) *cobra.Command {
+	cmd := &cobra.Command{
+		Use:               "aws CORE_INSTANCE",
+		Aliases:           []string{"ec2", "amazon"},
+		Short:             "Delete a core instance from Amazon EC2 (TODO)",
+		Args:              cobra.ExactArgs(1),
+		ValidArgsFunction: config.completeAggregators,
+		RunE: func(cmd *cobra.Command, args []string) error {
+			return errors.New("not implemented")
+		},
+	}
+	return cmd
+}
