+++ cmd/calyptia/delete_core_instance_gcp.go
@@ -1,0 +1,21 @@
+package main
+
+import (
+	"errors"
+
+	"github.com/spf13/cobra"
+)
+
+func newCmdDeleteCoreInstanceOnGCP(config *config) *cobra.Command {
+	cmd := &cobra.Command{
+		Use:               "gcp CORE_INSTANCE",
+		Aliases:           []string{"google", "gce"},
+		Short:             "Delete a core instance from Google Compute Engine (TODO)",
+		Args:              cobra.ExactArgs(1),
+		ValidArgsFunction: config.completeAggregators,
+		RunE: func(cmd *cobra.Command, args []string) error {
+			return errors.New("not implemented")
+		},
+	}
+	return cmd
+}
